impr: 完善代码逻辑
