+++ server/src/controller/performance.controller.ts
@@ -4,19 +4,13 @@ import { Optional } from 'sequelize/types';
 
 export async function createPerformance_c(ctx: Context) {
   const performanceInfo = ctx.request.body as Optional<any, string>;
+  const { aid } = performanceInfo;
   performanceInfo.ip = ctx.ip;
-  ctx.body = {
-    code: 200,
-    success: true,
-  };
-  try {
-    await createPerformance_s(performanceInfo);
 
-    ctx.defaultResponse({
-      code: 200,
-      message: 'success',
-      success: true,
-    });
+  try {
+    await createPerformance_s(aid, performanceInfo);
+
+    ctx.defaultResponse();
   } catch (err) {
     console.log(err);
 
@@ -26,14 +20,17 @@ export async function createPerformance_c(ctx: Context) {
 
 export async function getPerformanceData_c(ctx: Context) {
   const { type } = ctx.query;
+  const { aid } = ctx.state;
 
   try {
-    let data = await getPerformanceData_s(type as string);
+    const data = await getPerformanceData_s(aid, type as string);
     ctx.defaultResponse({
-      code: 200,
-      data: data,
+      data,
       message: '请求成功',
-      success: true,
     });
-  } catch (error) {}
+  } catch (err) {
+    console.log(err);
+
+    ctx.defaultError({ code: 500, message: '服务器出错' });
+  }
 }
