+++ server/src/middleware/auth.middleware.ts
@@ -1,9 +1,18 @@
 import jwt from 'jsonwebtoken';
 import env from '@/config/config.default';
 import type { Context, Next } from 'koa';
-import type { JwtPayload } from 'jsonwebtoken';
 
 const { JWT_SECRET } = env as Record<string, string>;
+const urls = [
+  '/api/v1/user/register',
+  '/api/v1/user/login',
+  '/api/v1/error/create',
+  '/api/v1/http/create',
+  '/api/v1/performance/create',
+  '/api/v1/resource/create',
+  '/api/v1/business/create',
+  '/api/v1/test',
+];
 
 export async function auth(ctx: Context, next: Next) {
   const {
@@ -11,29 +20,29 @@ export async function auth(ctx: Context, next: Next) {
     path,
   } = ctx.request;
 
-  if (path !== '/api/v1/user/register' && path !== '/api/v1/user/login') {
-    try {
-      const userInfo = jwt.verify(authorization, JWT_SECRET) as JwtPayload;
-
-      ctx.state.aid = userInfo.aid;
-    } catch (err: any) {
-      switch (err.name) {
-        case 'TokenExpiredError':
-          ctx.defaultError({
-            code: 400,
-            message: 'token已过期',
-          });
-          return;
-        case 'JsonWebTokenError':
-          ctx.defaultError({
-            code: 400,
-            message: '无效的token',
-          });
-          return;
-        default:
-          return;
+  if (!urls.includes(path)) {
+    jwt.verify(authorization, JWT_SECRET, (err, userInfo) => {
+      if (err) {
+        switch (err.name) {
+          case 'TokenExpiredError':
+            ctx.defaultError({
+              code: 400,
+              message: 'token已过期',
+            });
+            return;
+          case 'JsonWebTokenError':
+            ctx.defaultError({
+              code: 400,
+              message: '无效的token',
+            });
+            return;
+          default:
+            return;
+        }
       }
-    }
+
+      ctx.state.aid = (userInfo as jwt.JwtPayload).aid;
+    });
   }
 
   await next();
