+++ server/src/service/performance.service.ts
@@ -1,7 +1,21 @@
 import { PerformanceTimingModel, PerformancePaintModel, PerformanceCLSModel } from '@/model/performance.model';
+import UserModel from '@/model/user.model';
 import type { Optional } from 'sequelize/types';
 
-const timingType = ['DNS', 'TCP', 'SSL', 'FirstByte', 'TTFB', 'Trans', 'DomParse', 'Res', 'FP', 'TTI', 'DomReady', 'Load']
+const timingType = [
+  'DNS',
+  'TCP',
+  'SSL',
+  'FirstByte',
+  'TTFB',
+  'Trans',
+  'DomParse',
+  'Res',
+  'FP',
+  'TTI',
+  'DomReady',
+  'Load',
+];
 const typeDescribe: Record<string, string> = {
   DNS: 'DNS查询',
   TCP: 'TCP连接',
@@ -15,10 +29,13 @@ const typeDescribe: Record<string, string> = {
   TTI: '首次可交互时间',
   DomReady: 'DOM阶段渲染耗时',
   Load: '页面完全加载耗时',
-}
+};
 
+export async function createPerformance_s(aid: string, performanceInfo: Optional<any, string>) {
+  const isCreate = await findAid(aid);
 
-export async function createPerformance_s(performanceInfo: Optional<any, string>) {
+  if (!isCreate) return null;
+
   let result;
   if (performanceInfo.type === 'timing') {
     const {
@@ -36,7 +53,7 @@ export async function createPerformance_s(performanceInfo: Optional<any, string>
       bsVersion,
       ua,
       ip,
-    } = performanceInfo
+    } = performanceInfo;
     let data: Record<string, any> = {
       kind,
       type,
@@ -52,15 +69,15 @@ export async function createPerformance_s(performanceInfo: Optional<any, string>
       bsVersion,
       ua,
       ip,
-    }
+    };
     for (let type of timingType) {
-      data.timingType = type
-      data.describe = typeDescribe[type]
-      let [end, start] = performanceInfo[type].split('-')
-      data.during = Number(end) - Number(start)
-      data.start = Number(start)
-      data.end = Number(end)
-      await PerformanceTimingModel.create(data)
+      data.timingType = type;
+      data.describe = typeDescribe[type];
+      let [end, start] = performanceInfo[type].split('-');
+      data.during = Number(end) - Number(start);
+      data.start = Number(start);
+      data.end = Number(end);
+      await PerformanceTimingModel.create(data);
     }
   }
   if (performanceInfo.type === 'paint') {
@@ -72,7 +89,21 @@ export async function createPerformance_s(performanceInfo: Optional<any, string>
   return result;
 }
 
-export async function getPerformanceData_s(type: string) {
+/**
+ * @description: 查找用户表是否存在该aid
+ * @param aid 应用id
+ */
+export async function findAid(aid: string) {
+  const count = await UserModel.count({
+    where: {
+      aid,
+    },
+  });
+
+  return Boolean(count);
+}
+
+export async function getPerformanceData_s(aid: string, type: string) {
   let data;
   switch (type) {
     case 'timing':
@@ -82,11 +113,15 @@ export async function getPerformanceData_s(type: string) {
       data = PerformancePaintModel.findAll();
       break;
     case 'CLS':
-      data = PerformanceCLSModel.findAll()
+      data = PerformanceCLSModel.findAll();
       break;
     default:
-      data = PerformanceTimingModel.findAll({ limit: 20 })
+      data = PerformanceTimingModel.findAll({
+        where: {
+          aid,
+        },
+        limit: 20,
+      });
   }
-  return data
+  return data;
 }
-
