+++ server/src/service/user.service.ts
@@ -22,13 +22,10 @@ export async function userExistJudge(username: string) {
     },
   });
 
-  return !!count;
+  return Boolean(count);
 }
 
-function getToken(body: Record<string, any>) {
-  return jwt.sign(body, JWT_SECRET);
-}
-// 验证用户名密码
+// 验证用户名密码登陆
 export async function userPasswordJudge(username: string, password: string) {
   const res = await UserModel.findOne({
     where: {
@@ -37,17 +34,18 @@ export async function userPasswordJudge(username: string, password: string) {
     },
     attributes: ['username', 'aid'],
   });
-  if (res) {
-    const useInfo = res.get();
-    const token = getToken(useInfo);
 
-    return {
-      username: useInfo.username,
-      aid: useInfo.aid,
-      token,
-      permissions: ['/jsError', '/httpError', '/resourcesError', '/pageLoad', '/httpMonitor', '/test'],
-    };
-  } else {
-    return false;
-  }
+  return res ? getRefreshUserInfo(res.get()) : false;
+}
+
+export async function getRefreshUserInfo(oldUserInfo: any) {
+  const now = Date.now();
+  const token = jwt.sign(oldUserInfo, JWT_SECRET, { expiresIn: '2h' });
+
+  return {
+    ...oldUserInfo,
+    token,
+    permissions: ['/jsError', '/httpError', '/resourcesError', '/pageLoad', '/httpMonitor', '/test'],
+    expiresAt: now + 2 * 60 * 60 * 1000,
+  };
 }
