+++ src/monitor/ErrorVitals/errorClass.ts
