+++ src/monitor/PerformanceVitals/GetEntry.ts
