+++ src/monitor/global.d.ts
