+++ src/utils/constant/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @description: 一些常量名称
+ */
+
+// 未登录跳转路径
+export const notLoginPagePath = '/login';
+// 登录跳转路径
+export const loginPagePath = '/monitor';
+// localStorage存放的用户信息key
+export const userInfoKey = 'userInfo';
